Return usize from IDLInterface::get_prototype_depth.
It is used as an index into an array.

+++ components/script/dom/bindings/codegen/CodegenRust.py
@@ -1923,7 +1923,7 @@ impl IDLInterface for ${type} {
     fn get_prototype_id() -> PrototypeList::ID {
         PrototypeList::ID::${type}
     }
-    fn get_prototype_depth() -> uint {
+    fn get_prototype_depth() -> usize {
         ${depth}
     }
 }
